Vizsga eredmények

+++ sockets/exams/get/ProcessExamResults.js
@@ -6,12 +6,15 @@ module.exports = (socket) => {
 
     const removeFile = () => {
         const location = path.join(__dirname, `../../../temp_files/user_timer/${socket.handshake.session.cardNum}.json`)
-        const rawFileData = fs.readFileSync(location)
-        const time = Math.floor(((new Date().getTime() / 1000 - JSON.parse(rawFileData).timer.begin)))
+        if(fs.existsSync(location)){
+            const rawFileData = fs.readFileSync(location)
+            const time = Math.floor(((new Date().getTime() / 1000 - JSON.parse(rawFileData).timer.begin)))
         
-        fs.unlinkSync(location)
+            fs.unlinkSync(location)
 
-        return time
+            return time
+        }
+        return 0
     }
 
     const process = (answers, examCode) => {
@@ -27,7 +30,12 @@ module.exports = (socket) => {
         })
     }
 
-    socket.on('cancel-timer', removeFile)
+    socket.on('cancel-timer', () => {
+        const location = path.join(__dirname, `../../../temp_files/user_timer/${socket.handshake.session.cardNum}.json`)
+        if(fs.existsSync(location)){
+            fs.unlinkSync(location)
+        }
+    })
     
     socket.on('exam-finished', process)
 }
